refactor(observability): move controller metrics to dedicated package
