+++ internal/observability/metrics.go
@@ -1,4 +1,4 @@
-package controller
+package observability
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
