+++ internal/observability/metrics_test.go
@@ -1,4 +1,4 @@
-package controller
+package observability
 
 import (
 	"fmt"
